requirements.txt
discord.py>=2.3.0
flask>=2.3.0
gunicorn>=20.1.0
requests>=2.31.0